+++ src/core/orders/MidaBrokerOrder.ts
@@ -252,6 +252,10 @@ export class MidaBrokerOrder {
         return this._emitter.on(type, listener);
     }
 
+    public removeEventListener (uuid: string): void {
+        this._emitter.removeEventListener(uuid);
+    }
+
     private _notifyListeners (type: string, descriptor?: GenericObject): void {
         this._emitter.notifyListeners(type, descriptor);
     }